Avoid changing the size during realize
At the point when the the TerminalScreen is mapped, we are at a
funny state:

 - The terminal window is mapped
 - GTK+ hasn't yet applied the default size

Since the terminal window is mapped, we call gtk_window_resize()
instead of gtk_window_set_default_size() and end up mapping with
the wrong default size. (The observed effect is that depending on
timing, the terminal will be reset to 80x24 when switching to
a compositing window manager.)

The simple fix is to not resize the window when the TerminalScreen
is realized - changing sizing during realize is a poor idea in
all cases - but rather at the start of show().

http://bugzilla.gnome.org/show_bug.cgi?id=564648