Fix gridding for small sizes
The patch from 336325 (762e01947e) caused gridding to work wrong
when the number of rows/columns in the terminal got small enough
that other window elements were larger. Now that we have a better
fix for sizing of detached tabs (see previous commit), we can
revert the change.

The comment is made more explanatory to hopefully prevent the
code from accidentally being removed again in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=631709