help: rename gs-run-program to gs-execute-commands
Added steps.
Note about locating where a program is installed on the computer
is required because the user may try to execute a program not yet
installed on his computer, here the explaining to the user that
he/she can check if the program is installed on the computer by
using which enables them to successfully run a program.

For sake of clarity and that everything is a command in the CLI,
the page has been renamed more appropriately.

Page ID reflects the change in filename.