prefs: change UI strings in General tab
The checkbox labels should reflect that the mnemonics and the menu
accelerators are different and discrete from the keyboard shortcuts. At
the moment, the labels use non-standard terms to describe mnemonics and
menu accelerators, and the meanings are explained in the label as well.
Instead of using non-standard terms and an explanation, the label should
contain precise terms (as used in Gtk) and an explanation. There is no
harm in using the precise terms because these are advanced options. Bug
694740.