Remove the action and shortcut to set a static title from the menus
Setting a static title does not work because there is no code
listening to changes in TerminalScreen:description. The title is
dynamically set via escape sequences, usually from a suitable
PROMPT_COMMAND, which is noticed by VteTerminal.

If you want to customize the title of a particular terminal, please use
the escape sequence. This is what you had to use before, anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=724110