profile: Change the size UI according to latest designs
The check button to enable using a custom size has been replaced by a
reset button to restore the default size of 80x24. The labels were
tweaked and the entire row has been moved higher up in the dialog.

As a result of this, the 'use-custom-default-size' key is no longer
used and has been removed.

https://bugzilla.gnome.org/show_bug.cgi?id=724110