window: Attach popup menus to their parent widgets
When running on Wayland, popup menus must have a parent widget to attach
to. If a parent widget is not explicitly set, the GDK backend tries to
guess what parent widget seems appropriate. In order to have more
reliable popups, change the popups in the terminal window to set their
parent widgets explicitly.

https://bugzilla.gnome.org/show_bug.cgi?id=750755