window: Exit early from screen-removed handler for now empty notebook
When the notebook is now empty, it's not necessary to change anything else
since the window will be closed right away. Since GtkNotebook doesn't send
the switch-page signal when removing the last and only screen in the notebook,
priv->active_tab will point to an already freed object, which causes a
crash when the window tries to update itself for the new active tab.

https://bugzilla.gnome.org/show_bug.cgi?id=755240