profile: editor: Make colour buttons show their dialogue in editor mode
We use the colour buttons to edit the colour palette, so it's not good
usability to pop up the dialogue in palette mode. Since there is no way
to make the GtkColorButton's dialogue show up in editor mode, and since
GtkColorChooserDialog even unset editor mode on ::map, we need this gross
hack to work around it.

https://bugzilla.gnome.org/show_bug.cgi?id=756192