screen-container: Implement the resize popup for Wayland
Wayland doesn't have geometry hints for resizing. The size increments
are implemented entirely on the client side in GDK's Wayland backend.
This means that the WM cannot show the resize popup as it does on X.
Therefore, we have to implement it inside the application when running
on Wayland.

On X, the resize popup is shown when the mouse button is pressed near
the window edge and hidden when released. Unfortunately, this pointer
grab is entirely within the WM and not visible to the application.
Therefore, on Wayland, we show the resize popup when the size of the
window actually changes, and we hide on key-press-event or
motion-notify-event.

https://bugzilla.gnome.org/show_bug.cgi?id=760944