window: Make inspector available in non-debug builds
We disable the stock keybindings for the inspector because we want
all keys to go to the terminal instead of performing random actions,
and instead install a menu item in the Help menu to show the
inspector.

Previously this menu item was only present in a debug build. Since
gtk now has a settings schema that enables the inspector, use it to
show the menu item also in non-debug builds when the key is enabled.

Since this is a foreign settings schema that could be considered to
be internal to gtk+, be extra careful in instantiating it only if
it exists, and the expected key exists with the expected value type.

https://bugzilla.gnome.org/show_bug.cgi?id=770245