window: Fix CSD size calculations with long titles
To get the size of the window use the actual allocation rather then the
preferred size as the latter takes into account the natural size of the
title bar as well as the content which throws of calculation when the
natural size of the title bar is wider then the content.

Also use the actual allocation of the content specifically for the
calculation of the CSD size such that both are in sync, but keep using
the preferred size otherwise as those values get used for resizing
calculations.

https://bugzilla.gnome.org/show_bug.cgi?id=769898
(cherry picked from commit cae04637f696e30f9054482411eb0fabd6064a84)