all: Add support for OSC 8 hyperlinks (HTML-like anchors)
The escape sequences
  OSC 8 ; params ; URI BEL
  OSC 8 ; params ; URI ST
turn subsequent characters into an HTML-like anchor to the given address.
Terminate the hyperlink with the same escape sequence with an empty URI.

Params is an optional colon-separated list of key=value assignments. Cells
having the same URI and "id" parameter, or cells lacking an "id" that were
printed in a single OSC 8 run are underlined together on mouseover.

For further details, see
https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda.

https://bugzilla.gnome.org/show_bug.cgi?id=779734