nautilus: Use the default display
When creating a new terminal window, use the default display
instead of the display/screen that the nautilus window is on.

For some reason, the nautilus process that draws the desktop
background is on a GdkX11Display, which made the terminal
window created be on X/Xwayland instead of wayland. This caused
the terminal window not to have the correct theme and
clipboard connection.

https://bugs.launchpad.net/ubuntu/+source/gnome-terminal/+bug/1722121