client: legacy: Improve --tab behaviour to open a new tab in an existing window
An often-requested feature has been for "gnome-terminal --tab" to open a
new tab in an existing window, see the discussion in bug 83203. While
in general, when called from an external source, it's often unclear
which window to choose, we can at least provide this when called from
within gnome-terminal.

For this, introduce some environment variables GNOME_TERMINAL_SERVICE
that is the D-Bus unique name of the gnome-terminal-server instance for
the terminal, and GNOME_TERMINAL_SCREEN that is the D-Bus object path
for the D-Bus controller for that terminal.

On the server side, these variables will be set for each terminal
created in the environment of the child process.

On the client side, these variables will be used to define which
window the "--tab" option refers to. Also, optionally, when using
the new "--print-environment" option, the client will print the
corresponding variables for the *newly created* terminal.

https://bugzilla.gnome.org/show_bug.cgi?id=83203