all: Improve mnemonic handling
Disable mnemonics for the submenus (Alt+F for File etc.) when the menubar
is shown by the shell (e.g. Unity) so that they are usable in the terminal.
(As of the GMenu port, disabling them via the config option didn't work.)
Also hide the corresponding preference checkbox.

Make mnemonics work in the Find and Preferences windows.

https://bugzilla.gnome.org/show_bug.cgi?id=792978