app: Revise the logic for having an appmenu
Display the appmenu as the 4th window decoration button in GNOME Shell,
when the shell doesn't show the appmenu. This was accidentally gone with
the GMenu port.

Don't have an appmenu in Unity where the shell displays both the appmenu
and the menubar, the appmenu is pointless in this case.

https://bugzilla.gnome.org/show_bug.cgi?id=793135