window: Expose find action in headerbar
Find is a useful standard action, and now that we are using client-side
decorations, we have a place to expose it more prominently than hiding
it in the menu.

For now this simply opens the existing dialog, in the future it may be
worth considering to make the UI available in a popover as well.

https://bugzilla.gnome.org/show_bug.cgi?id=756798