prefs: Hide default-menubar setting when headerbars are used
As headerbars combine the functionality of titlebar, toolbar and
menubar, they generally replace (not supplement) those elements.

For that reasons, the default-show-menubar setting is ignored when
headerbars are enabled. However despite not having any effect, the
corresponding checkbox is still shown in the preferences. It doesn't
make sense to expose non-working settings, so hide it when headerbars
are used.

https://gitlab.gnome.org/GNOME/gnome-terminal/issues/81