window: Ignore resize if fullscreen
The terminal update size function will not change the window size if
maximized or tiled, but does not account for fullscreen window.

As a result, if the terminal window is fullscreen, the size is updated
which cause the window to shrink because of size increments.

Make sure `terminal_window_update_size()` accounts for fullscreen state
as well and return early in that case.

https://gitlab.gnome.org/GNOME/gnome-terminal/issues/158
(cherry picked from commit b0d5d4a83fe223d419b824b191edaf05eb96b533)