systemd: Add systemd helper to move processes into scopes
Doing this creates a better separation between the server and the
different terminals.

This helper is basically copied from gnome-desktop. Copying the code is
not ideal, but not a lot of GNOME applications have a use for this and
it is not clear how a more generic API should looks like.

Luckily this change is rather self contained and a port to another API
will be trivial when it becomes available.