build: Stop terminal service together with graphical session
There is no point in gnome-terminal-server to survive the graphical
session. It will generally die anyway because the connection to the
display server is lost, but it does not hurt to make this explicit.

Also add a stop timeout to ensure the server gets killed if it refuses
to shut down.