build: Bind terminal children to graphical session
This ensures that all processes launched by gnome-terminal (and their
children) are reaped when the user logs out of the graphical session.

Note that this moves the killing of lingering processes from the point
where the systemd user instance is shut down to the the point where the
user is logging out of the graphical session.

A slightly more conservative approach might be to keep setting the
KillMode. However, as said above, this will just delay the reaping by 10
seconds. If processes need/want to linger, they will need to deal with
systemd properly to achieve this. So really, we are "just" increasing
the pressure on projects to adjust to a systemd managed worled.

Note that this setup does still permit processes to survive a
gnome-terminal-server crash or stop.