build: Group terminal processes into a slice
This effectively changes the cgroup hierarchy making all terminal window
scope units and the main server unit be sibblings of
apps-org.gnome.Terminal.slice rather than being in the root slice.

Note that using the apps prefix with the reverse domain name has not yet
been standardised on. But it is very likely that this will be the
standard that everything will converge on in the future.

https://gitlab.gnome.org/GNOME/gnome-terminal/issues/206