profile: New default colour scheme based on the GNOME 3.32 icon palette
Also use these colours for foreground and background, when not using theme colours.

I chose colours from the palette to provide good contrast in a wide
variety of applications on both light and dark backgrounds, taking into
account that the "Show bold text in bright colors" option is now off by
default.

The explanation behind some of the colour choices:

* I chose pure white for slot 15, since it works better as a light
  terminal background colour than grey (better contrast vs green, cyan)
* To improve contrast in blue-background applications, I chose a darker
  blue along with a brighter green and cyan. The dark blue desired was
  darker than available on the palette, so I made a darker blue by
  blending Blue 5 with Dark 4.
* To provide extra personality, reference the classic CGA/VGA palette,
  and improve readability on light backgrounds, I picked a middle Brown
  tone for slot 3 rather than make up a muddy "dark yellow".

Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/issues/72