client: legacy: Suppress debug message spam
The glib log writer API has a deficiency in that the filtering is
done in the default log writer, instead of only passing messages
that pass the filter. This is filed as glib#2087, but until that
is fixed, apply a simple log level filter to work around it.

Patch by Kim Nguyen <https://gitlab.gnome.org/kim.nguyen>

Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/42