server: Fix refcount leak
Ownership of the GDBusMethodInvocation is passed to terminal_receiver_impl_exec,
and will be released by terminal_receiver_complete_exec. The extra reference
added to ExecData was leaked, causing the GUnixFDList passed from the client
(if any) also to be leaked, which meant that its file descriptors were never
closed, leading to FD exhaustion.