Revert "screen: Use clean env when creating new tab"
The patch is correct, but it exposes a deficiency in the desktop
environment in that the WM/shell's environment contains necessary
env vars (e.g. ssh-agent variables) that are missing from the
systemd --user / d-bus activation environment.  The desktop will
need to update the activation environment, but until that is done,
reverting the patch will revert gnome-terminal to the previous
behaviour of effectively using the environment of the client
of the first terminal created for all new terminals opened via
new terminal/tab/window.

This reverts commit 9bb94e3aab84ecc4e7733d0ee001ee50256bd273.

https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/253