man: Document the various ways to run a command in a gnome-terminal
There are three ways to run a command in a gnome-terminal (using
"python -q" as my example because it illustrates how argument parsing
works), of which "-e" was previously the only one listed in the
man page:

* gnome-terminal [OPTIONS] -- python3 -q
* gnome-terminal [OPTIONS] -e "python3 -q" [MORE OPTIONS]
* gnome-terminal [OPTIONS] -x python3 -q

--command is a long form of -e, and --execute is a long form of -x.
Both the "-e" and "-x" forms are deprecated in favour of the "--" form,
but previously only the deprecated "-e" form was documented in the
man page.

Document the preferred form in the synopsis, and document -e and -x in
terms of their replacements with the preferred form.

Resolves: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/296