accels: Make it convenient to use the '+' alphanumeric key
Currently, the default accelerator for zooming in (ie., Ctrl++)
requires pressing the shift key when using the alphanumeric keys on
keyboards where the '+' is on the upper level. This is inconvenient.
Even more so on laptop keyboards where the numeric keypad isn't easily
accessible.

From now on, any accelerator that has the '+' key will work without
requiring the shift key.

https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/313