all: Add settings schema verifier
Add external schemas, and build a reference gschema.compiled. At
startup, verify that the installed compiled schemas conform to that
reference schema; if not, use the reference schema as schema source
instead.

This fixes the problem where, on upgrade, or through an otherwise failed
recompile of the gschemas.compiled cache, where the upgrade to
gnome-terminal added ettings keys, the server will abort when trying to get
that new key.

E.g. https://bugzilla.redhat.com/show_bug.cgi?id=1961775 which is just
the latest instance of this problem occurring.