all: backport: Implement xdg-terminal-exec support
Add support to read and write the default terminal as determined by
xdg-terminal-exec, as well as UI to make gnome-terminal the default
terminal.

This bumps the meson req version to 0.61.0 (for install_symlink()).

This is required on the stable branch to get ahead of the glib
changes in glib issue 338 and associated merge requests.

Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/7942
(cherry picked from commit f1814cb7ecfe6a5cb4f629dcea56db0dec10a242)