server: Add theme style support
Uses libhandy to monitor the xdg-desktop-portal color-scheme
property and switch themes accordingly. This allows Terminal to
switch colors when the GNOME Dark Style option is toggled.
If not compiled with libhandy, or if the portal color-scheme is
not supported, fall back to using the old dark theme system.

Based on a patch by "robxnano".

Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/7972
Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/190