prefs: Rework theme colours option
The API to get the actual theme colours is deprecated, and appears to
have stopped working correcly, see
https://bugzilla.redhat.com/show_bug.cgi?id=2237631 .

Instead, use the HdyStyleManager's view of whether the theme is dark or
light, and use the palette colours themselves to set the fg and bg
colours.  This is also similar to what gnome-console does.

Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/8000