notebook: port notebook to AdwTabView
This also moves the AdwTabBar to the window which is much more convenient
for how it gets used by the TerminalWindow. The end action area menu is
still maintained to provide the list of all tabs.

Now that we use AdwTabView, we need to confirm success/failure when
closing a tab. This adds the mechanism to do that and connects it into
the TerminalWindow workflow.