screen: show terminal size during resizes
This is added to the screen so that we can place the label in the
correct position whether or not a GtkScrolledWindow is used or a
GtkScrollbar which is external.

We may want to ignore tiled states as well, but this is a good start.

Similar things will need to be done for context menus so the additional
class overrides will be reused for that too.

The TerminalScreen now uses screen.ui for various setup and styling so
we can define the subwidgets in the template.