window: port window geometry to GTK 4
This is still going to heavily rely on fixing a bug in GdkWaylandToplevel
where gtk_window_set_default_size() is not yet doing the right thing. It
is supposed to cause the toplevel to resize when you set to -1,-1 taking
the new size-requests into account. See GNOME/gtk#6157 for the underlying
issue there.

Since GeometryHints are gone and not coming back, this tries to get the
rest of things in place to deal with that. If you don't manually resize the
window, you should see resizes happening correctly from the Advanced menu
by selecting a new "columns × rows" item.