screen: implement various DnD formats
This adds the portal drop targets as well as ensuring that we do proper
fallback to text/uri-list when we fail to get the GDK_TYPE_FILE_LIST. This
can happen when we get a drop from a portal-enabled application where
something like a directory was dropped such as Nautilus.

Additionally support for text/x-moz-url is retained but now uses a similar
code path to text/uri-list. Since we need to do UCS-2 to UTF-8 conversion
and specialized line handling, we read a line at a time asynchronously all
the same. The GConverter and GDataInputStream handle those quirks for us
transparently.