screen: add DnD drop highlighting
This uses a AdwBin as a child widget (it could really be any type of
widget though) which is snapshot when the drop target is active. It just
draws a 1px border using box-shadow to avoid cascading relayout.

Since we have padding on the TerminalScreen this shows up on the outer
most pixel of the widget. We want this because otherwise the rectangle
cursor can draw beyond the extent of the highlight and that looks strange.