util: check flatpak for alternate terminals
If the alternate default terminals are installed via Flatpak, then the
code would previously always overwrite the default terminal with GNOME
Terminal.

This checks the user and system installation to see if the terminal is
installed there. We could do this better if we linked against libflatpak
to get all of the System installations, but that is probably more effort
than it is worth.

(cherry picked from commit 5ac3b8e4bd6fa02651b3c23cedb0a7e1cd769655)