window: Revert to previous tab insertion behaviour
Fixes a regression from 0f9a83fa4cddaf703662d0eecad1c086b12f647a.

The AdwTabView behaviour when used with adw_tab_view_add_page() and
a parent tab is too confusing to be useful.  Instead, revert to
inserting the new tab immediately to the right of the parent tab,
or, if the parent tab is pinned, immediately after the last pinned
tab.

https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/8025